+++ packages/gov-ui-kit/CHANGELOG.md
@@ -14,7 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
--   Implement `InputDate` component
+-   Implement `InputDate` and `Avatar` components
 
 ## [1.0.8] - 2024-01-17